+++ aries-site/src/examples/templates/filtering/FilteringTable/index.js
@@ -1,1 +1,2 @@
 export * from './FilteringTable';
+export * from './FilterServers';
