+++ aries-site/src/pages/components/pageheader.mdx
@@ -137,6 +137,7 @@ Child pages include a reverse anchor which allows the user to navigate from the 
   code={[
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/components/pageheader/ChildPageHeaderExample.js',
     'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/ReverseAnchor/ReverseAnchor.js',
+    'https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/templates/filtering/FilteringTable/FilterServers.js',
   ]}
   docs="https://v2.grommet.io/pageheader?theme=hpe"
   github="https://github.com/grommet/hpe-design-system/tree/master/aries-site/src/examples/components/pageheader/ChildPageHeaderExample.js"
@@ -153,6 +154,7 @@ Top-level pages do not display a reverse anchor because they do not have a paren
 <Example
   code="https://raw.githubusercontent.com/grommet/hpe-design-system/master/aries-site/src/examples/components/pageheader/TopLevelPageHeaderExample.js"
   docs="https://v2.grommet.io/pageheader?theme=hpe"
+  github="https://github.com/grommet/hpe-design-system/tree/master/aries-site/src/examples/components/pageheader/TopLevelPageHeaderExample.js"
   showResponsiveControls={['fullScreen']}
   screenContainer={{ scale: 0.6 }}
 >
